repackage helper.manager to helper.portals
