+++ src/main/java/com/rebuild/server/business/charts/ChartData.java
@@ -30,8 +30,8 @@ import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONArray;
 import com.alibaba.fastjson.JSONObject;
 import com.rebuild.server.Application;
-import com.rebuild.server.helper.manager.PickListManager;
-import com.rebuild.server.helper.manager.value.FieldValueWrapper;
+import com.rebuild.server.helper.portals.PickListManager;
+import com.rebuild.server.helper.portals.value.FieldValueWrapper;
 import com.rebuild.server.metadata.EntityHelper;
 import com.rebuild.server.metadata.MetadataHelper;
 import com.rebuild.server.metadata.entityhub.DisplayType;
