+++ src/main/java/com/rebuild/server/business/datalist/DataWrapper.java
@@ -24,7 +24,7 @@ import org.apache.commons.logging.LogFactory;
 
 import com.alibaba.fastjson.JSON;
 import com.rebuild.server.Application;
-import com.rebuild.server.helper.manager.value.FieldValueWrapper;
+import com.rebuild.server.helper.portals.value.FieldValueWrapper;
 import com.rebuild.server.metadata.MetadataHelper;
 import com.rebuild.server.metadata.entityhub.EasyMeta;
 import com.rebuild.utils.JSONUtils;
