+++ src/main/java/com/rebuild/server/business/datalist/JSONQueryParser.java
@@ -27,7 +27,7 @@ import org.apache.commons.lang.math.NumberUtils;
 
 import com.alibaba.fastjson.JSONArray;
 import com.alibaba.fastjson.JSONObject;
-import com.rebuild.server.helper.manager.AdvFilterManager;
+import com.rebuild.server.helper.portals.AdvFilterManager;
 import com.rebuild.server.metadata.EntityHelper;
 import com.rebuild.server.metadata.MetadataHelper;
 import com.rebuild.server.metadata.entityhub.DisplayType;
