--- src/main/java/com/rebuild/server/helper/manager/PortalsManager.java
@@ -1,28 +1,0 @@
-/*
-rebuild - Building your business-systems freely.
-Copyright (C) 2018 devezhao <zhaofang123@gmail.com>
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program. If not, see <https://www.gnu.org/licenses/>.
-*/
-
-package com.rebuild.server.helper.manager;
-
-/**
- * 前端配置管理器。标记接口
- * 
- * @author devezhao
- * @since 01/04/2019
- */
-public interface PortalsManager {
-}
