+++ src/main/java/com/rebuild/server/helper/portals/LayoutManager.java
@@ -16,7 +16,7 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <https://www.gnu.org/licenses/>.
 */
 
-package com.rebuild.server.helper.manager;
+package com.rebuild.server.helper.portals;
 
 import com.alibaba.fastjson.JSON;
 import com.rebuild.server.Application;
