+++ src/main/java/com/rebuild/server/helper/portals/SharableManager.java
@@ -16,7 +16,7 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <https://www.gnu.org/licenses/>.
 */
 
-package com.rebuild.server.helper.manager;
+package com.rebuild.server.helper.portals;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
