+++ src/main/java/com/rebuild/server/service/base/QuickCodeReindexTask.java
@@ -25,7 +25,7 @@ import org.apache.commons.lang.math.NumberUtils;
 
 import com.github.stuxuhai.jpinyin.PinyinHelper;
 import com.rebuild.server.Application;
-import com.rebuild.server.helper.manager.PickListManager;
+import com.rebuild.server.helper.portals.PickListManager;
 import com.rebuild.server.helper.task.BulkTask;
 import com.rebuild.server.metadata.EntityHelper;
 import com.rebuild.server.metadata.entityhub.DisplayType;
