+++ src/main/java/com/rebuild/web/admin/entityhub/PickListControll.java
@@ -29,7 +29,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import com.alibaba.fastjson.JSONArray;
 import com.alibaba.fastjson.JSONObject;
 import com.rebuild.server.Application;
-import com.rebuild.server.helper.manager.PickListManager;
+import com.rebuild.server.helper.portals.PickListManager;
 import com.rebuild.server.metadata.MetadataHelper;
 import com.rebuild.server.metadata.entityhub.PickListService;
 import com.rebuild.web.BaseControll;
