+++ src/main/java/com/rebuild/web/admin/entityhub/ViewAddonsControll.java
@@ -32,7 +32,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 
 import com.alibaba.fastjson.JSON;
 import com.rebuild.server.Application;
-import com.rebuild.server.helper.manager.ViewAddonsManager;
+import com.rebuild.server.helper.portals.ViewAddonsManager;
 import com.rebuild.server.metadata.EntityHelper;
 import com.rebuild.server.metadata.MetadataHelper;
 import com.rebuild.server.metadata.entityhub.EasyMeta;
