+++ src/main/java/com/rebuild/web/base/entity/AdvFilterControll.java
@@ -31,8 +31,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;
 import com.rebuild.server.Application;
-import com.rebuild.server.helper.manager.AdvFilterManager;
-import com.rebuild.server.helper.manager.SharableManager;
+import com.rebuild.server.helper.portals.AdvFilterManager;
+import com.rebuild.server.helper.portals.SharableManager;
 import com.rebuild.server.metadata.EntityHelper;
 import com.rebuild.server.service.bizz.UserHelper;
 import com.rebuild.server.service.query.AdvFilterParser;
