+++ src/main/java/com/rebuild/web/base/entity/DataListSettingsControll.java
@@ -36,9 +36,9 @@ import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONArray;
 import com.alibaba.fastjson.JSONObject;
 import com.rebuild.server.Application;
-import com.rebuild.server.helper.manager.DataListManager;
-import com.rebuild.server.helper.manager.LayoutManager;
-import com.rebuild.server.helper.manager.SharableManager;
+import com.rebuild.server.helper.portals.DataListManager;
+import com.rebuild.server.helper.portals.LayoutManager;
+import com.rebuild.server.helper.portals.SharableManager;
 import com.rebuild.server.metadata.EntityHelper;
 import com.rebuild.server.metadata.MetadataHelper;
 import com.rebuild.server.metadata.MetadataSorter;
