+++ src/main/java/com/rebuild/web/base/entity/GeneralDataListControll.java
@@ -32,7 +32,7 @@ import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;
 import com.rebuild.server.business.datalist.DataList;
 import com.rebuild.server.business.datalist.DefaultDataList;
-import com.rebuild.server.helper.manager.DataListManager;
+import com.rebuild.server.helper.portals.DataListManager;
 import com.rebuild.server.metadata.MetadataHelper;
 import com.rebuild.web.BaseEntityControll;
