+++ src/main/java/com/rebuild/web/base/entity/ReferenceSearch.java
@@ -35,7 +35,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.rebuild.server.Application;
-import com.rebuild.server.helper.manager.value.FieldValueWrapper;
+import com.rebuild.server.helper.portals.value.FieldValueWrapper;
 import com.rebuild.server.metadata.EntityHelper;
 import com.rebuild.server.metadata.MetadataHelper;
 import com.rebuild.server.service.bizz.UserHelper;
