+++ src/test/java/com/rebuild/server/helper/portals/FormsManagerTest.java
@@ -1,6 +1,6 @@
 /*
-rebuild - Building your system freely.
-Copyright (C) 2019 devezhao <zhaofang123@gmail.com>
+rebuild - Building your business-systems freely.
+Copyright (C) 2018 devezhao <zhaofang123@gmail.com>
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -16,23 +16,34 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <https://www.gnu.org/licenses/>.
 */
 
-package com.rebuild.server.helper.manager;
+package com.rebuild.server.helper.portals;
 
 import org.junit.Test;
 
 import com.alibaba.fastjson.JSON;
 import com.rebuild.server.TestSupport;
+import com.rebuild.server.helper.portals.FormsManager;
 import com.rebuild.server.service.bizz.UserService;
 
 /**
- * @author devezhao zhaofang123@gmail.com
- * @since 2019/03/09
+ * 
+ * @author devezhao
+ * @since 01/04/2019
  */
-public class DashboardManagerTest extends TestSupport {
+public class FormsManagerTest extends TestSupport {
 
 	@Test
-	public void testGetList() throws Exception {
-		JSON dashs = DashboardManager.getDashList(UserService.ADMIN_USER);
-		System.out.println(dashs.toJSONString());
+	public void testModel() throws Exception {
+		JSON newModel = FormsManager.getFormModel("User", UserService.ADMIN_USER);
+		System.out.println(newModel);
+		
+		JSON editModel = FormsManager.getFormModel("User", UserService.ADMIN_USER, UserService.SYSTEM_USER);
+		System.out.println(editModel);
+	}
+	
+	@Test
+	public void testViewModel() throws Exception {
+		JSON viewModel = FormsManager.getViewModel("User", UserService.ADMIN_USER, UserService.SYSTEM_USER);
+		System.out.println(viewModel);
 	}
 }
